PEN-18 refactoring after dev review
